Refactored tests into different packages
