+++ src/test/java/com/ibm/stocator/fs/swift2d/systemtests/SwiftBaseTest.java
@@ -15,7 +15,7 @@
  *
  */
 
-package com.ibm.stocator.fs.swift2d;
+package com.ibm.stocator.fs.swift2d.systemtests;
 
 import java.io.IOException;
 import java.net.URI;
