+++ src/test/java/com/ibm/stocator/fs/swift2d/systemtests/SwiftFileSystemBaseTest.java
@@ -16,15 +16,13 @@
  *  limitations under the License.
  */
 
-package com.ibm.stocator.fs.swift2d;
+package com.ibm.stocator.fs.swift2d.systemtests;
 
 import java.io.IOException;
 
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.Path;
 import org.junit.Assume;
-
-import static com.ibm.stocator.fs.swift2d.SwiftTestUtils.noteAction;
 
 /**
  * This is the base class for most of the Swift tests
@@ -35,10 +33,10 @@ public class SwiftFileSystemBaseTest extends SwiftBaseTest {
 
   @Override
   public void setUp() throws Exception {
-    noteAction("setup");
+    SwiftTestUtils.noteAction("setup");
     super.setUp();
     Assume.assumeNotNull(getFs());
-    noteAction("setup complete");
+    SwiftTestUtils.noteAction("setup complete");
   }
 
   /**
@@ -48,7 +46,7 @@ public class SwiftFileSystemBaseTest extends SwiftBaseTest {
    * @param description test description
    */
   protected void describe(String description) {
-    noteAction(description);
+    SwiftTestUtils.noteAction(description);
   }
 
   /**
