+++ src/test/java/com/ibm/stocator/fs/swift2d/systemtests/SwiftTestUtils.java
@@ -15,7 +15,7 @@
  *
  */
 
-package com.ibm.stocator.fs.swift2d;
+package com.ibm.stocator.fs.swift2d.systemtests;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
