+++ src/test/java/com/ibm/stocator/fs/swift2d/systemtests/TestSwiftOperations.java
@@ -15,7 +15,7 @@
  *
  */
 
-package  com.ibm.stocator.fs.swift2d;
+package com.ibm.stocator.fs.swift2d.systemtests;
 
 import java.io.IOException;
 import java.text.MessageFormat;
@@ -24,7 +24,6 @@ import com.ibm.stocator.fs.common.ObjectStoreGlobFilter;
 import com.ibm.stocator.fs.common.ObjectStoreGlobber;
 import org.apache.hadoop.fs.FileStatus;
 import org.apache.hadoop.fs.Path;
-import org.junit.Assert;
 import org.junit.Assume;
 import org.junit.Before;
 import org.junit.Test;
@@ -60,7 +59,7 @@ public class TestSwiftOperations extends SwiftBaseTest {
     createEmptyFile(new Path(getBaseURI(),
         MessageFormat.format(sparkSuccessFormat, new Object[]{objectName})));
     FileStatus[] stats = sFileSystem.listStatus(new Path(getBaseURI() + "/" + objectName));
-    Assert.assertEquals(11, stats.length);
+    assertEquals(11, stats.length);
     // read 11 objects
     for (int i = 0;i < 11; i++) {
       String id = String.format("%0" + 2 + "d", i);
@@ -68,7 +67,7 @@ public class TestSwiftOperations extends SwiftBaseTest {
       Path path = new Path(getBaseURI(), MessageFormat.format(swiftDataFormat, params));
       byte[] res = SwiftTestUtils.readDataset(getFs(),
           path, data.length);
-      Assert.assertArrayEquals(data, res);
+      assertArrayEquals(data, res);
     }
     // delete 11 objects
     for (int i = 0;i < 11; i++) {
@@ -81,17 +80,17 @@ public class TestSwiftOperations extends SwiftBaseTest {
     getFs().delete(new Path(getBaseURI(),
             MessageFormat.format(sparkSuccessFormat, new Object[]{objectName})), false);
     stats = getFs().listStatus(new Path(getBaseURI() + "/" + objectName));
-    Assert.assertEquals(0, stats.length);
+    assertEquals(0, stats.length);
   }
 
   @Test
   public void testFileExists() throws IOException {
     Path testFile = new Path(getBaseURI() + "/testFile");
     createFile(testFile, data);
-    Assert.assertTrue(getFs().exists(testFile));
+    assertTrue(getFs().exists(testFile));
     getFs().delete(testFile, false);
     FileStatus[]  stats = getFs().listStatus(testFile);
-    Assert.assertEquals(0, stats.length);
+    assertEquals(0, stats.length);
   }
 
   @Test
@@ -101,12 +100,12 @@ public class TestSwiftOperations extends SwiftBaseTest {
       createFile(new Path(getBaseURI() + name), data);
     }
     FileStatus[] results = getFs().globStatus(new Path(getBaseURI() + "/File*"));
-    Assert.assertEquals(2, results.length);
+    assertEquals(2, results.length);
     for (String name : testFileNames) {
       getFs().delete(new Path(getBaseURI() + name), false);
     }
     results = getFs().listStatus(new Path(getBaseURI()));
-    Assert.assertEquals(0, results.length);
+    assertEquals(0, results.length);
   }
 
   @Test
@@ -121,32 +120,32 @@ public class TestSwiftOperations extends SwiftBaseTest {
     ObjectStoreGlobber globber = new ObjectStoreGlobber(getFs(), wildcard,
             new ObjectStoreGlobFilter(""));
     FileStatus[] results = globber.glob();
-    Assert.assertEquals(3, results.length);
+    assertEquals(3, results.length);
 
     wildcard = new Path(getBaseURI() + "/Dir/*"); // Files in "Dir" directory
     globber = new ObjectStoreGlobber(getFs(), wildcard, new ObjectStoreGlobFilter(""));
     results = globber.glob();
-    Assert.assertEquals(3, results.length);
+    assertEquals(3, results.length);
 
     wildcard = new Path(getBaseURI() + "/Dir/SubDir/*"); // Files in "SubDir" directory
     globber = new ObjectStoreGlobber(getFs(), wildcard, new ObjectStoreGlobFilter(""));
     results = globber.glob();
-    Assert.assertEquals(2, results.length);
+    assertEquals(2, results.length);
 
     wildcard = new Path(getBaseURI() + "/*1"); // Files ending in "1"
     globber = new ObjectStoreGlobber(getFs(), wildcard, new ObjectStoreGlobFilter(""));
     results = globber.glob();
-    Assert.assertEquals(2, results.length);
+    assertEquals(2, results.length);
 
     wildcard = new Path(getBaseURI() + "/Dir/SubDir/*2"); // Files in "SubDir" ending with "2"
     globber = new ObjectStoreGlobber(getFs(), wildcard, new ObjectStoreGlobFilter(""));
     results = globber.glob();
-    Assert.assertEquals(1, results.length);
+    assertEquals(1, results.length);
 
     wildcard = new Path(getBaseURI() + "/Dir/*/File1"); // Files called File1 in a SubDir
     globber = new ObjectStoreGlobber(getFs(), wildcard, new ObjectStoreGlobFilter(""));
     results = globber.glob();
-    Assert.assertEquals(1, results.length);
+    assertEquals(1, results.length);
 
   }
 }
